fix: fixed plugin install method
now the plugin install method is only one, and it expects that the name of the plugin to be installed is passed as parameter. This requires some fixes also on the components using this method.
